feat(IIH-40): questions set changes
