+++ apps/client/src/app/(protected)/(dashboard)/interviewer/interviews/interview/page.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const InterviewPage = () => (
+  <div>
+    <div>Interview Page!</div>
+  </div>
+);
+
+export default InterviewPage;
