+++ apps/client/src/providers/store/selectors/intervew.ts
@@ -1,0 +1,3 @@
+import { TStateModel } from '..';
+
+export const getInterviewId = (state: TStateModel) => state?.interview.intreviewId;
