+++ apps/client/src/providers/store/selectors/uesr.ts
@@ -1,0 +1,3 @@
+import { TStateModel } from '..';
+
+export const getUserData = (state: TStateModel) => state?.users.user;
